requirements.txt
# Core deep-learning stack (Updated for Python 3.12)
torch==2.2.1
torchvision==0.17.1

# Diffusion & ControlNet
diffusers==0.25.1
transformers==4.36.2
accelerate==0.26.1
xformers==0.0.24

# MediaPipe & Protobuf (Specific pins for stability)
protobuf==4.25.3
mediapipe==0.10.14
controlnet-aux==0.0.7

# Computer vision & image processing
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
onnxruntime==1.17.1

# HuggingFace & Utilities
huggingface-hub==0.20.3
matplotlib==3.8.2
scipy==1.12.0
einops==0.7.0
omegaconf==2.3.0
safetensors==0.4.2
filelock==3.13.1
packaging==23.2
requests==2.31.0
tqdm==4.66.2